+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- 添加遗漏的 getClip 函数 [#6](https://github.com/batu1579/hamibot-types/issues/6)
+
 ## [0.1.3] - 2022-12-04
 
 ### Fixed